python/requirements.txt
openai
numpy
scikit-learn
PyPDF2
python-dotenv